# backend/src/main/java/com/product/template/controller/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.12.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.product.template.controller;

import com.product.template.dto.ErrorResponse;
import com.product.template.dto.PageUserResponse;
import com.product.template.dto.UpdateUserRequest;
import com.product.template.dto.UserResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.12.0")
@Validated
@Tag(name = "Users", description = "ユーザー管理API")
public interface UsersApi {

    /**
     * DELETE /api/users/{id} : ユーザー削除（管理者専用）
     * 指定したIDのユーザーを削除する。管理者権限が必要。
     *
     * @param id ユーザーID (required)
     * @return 削除成功 (status code 204)
     *         or 未認証 (status code 401)
     *         or 権限不足（管理者権限が必要） (status code 403)
     *         or ユーザーが見つからない (status code 404)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "ユーザー削除（管理者専用）",
        description = "指定したIDのユーザーを削除する。管理者権限が必要。",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "204", description = "削除成功"),
            @ApiResponse(responseCode = "401", description = "未認証", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "権限不足（管理者権限が必要）", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "ユーザーが見つからない", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/users/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteUser(
        @Parameter(name = "id", description = "ユーザーID", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /api/users/me : ログインユーザー情報取得
     * 認証済みユーザー自身の情報を取得する
     *
     * @return 取得成功 (status code 200)
     *         or 未認証 (status code 401)
     */
    @Operation(
        operationId = "getCurrentUser",
        summary = "ログインユーザー情報取得",
        description = "認証済みユーザー自身の情報を取得する",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "取得成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "未認証", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/users/me",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserResponse> getCurrentUser(
        
    );


    /**
     * GET /api/users/{id} : ユーザー詳細取得（管理者専用）
     * 指定したIDのユーザー情報を取得する。管理者権限が必要。
     *
     * @param id ユーザーID (required)
     * @return 取得成功 (status code 200)
     *         or 未認証 (status code 401)
     *         or 権限不足（管理者権限が必要） (status code 403)
     *         or ユーザーが見つからない (status code 404)
     */
    @Operation(
        operationId = "getUserById",
        summary = "ユーザー詳細取得（管理者専用）",
        description = "指定したIDのユーザー情報を取得する。管理者権限が必要。",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "取得成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "未認証", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "権限不足（管理者権限が必要）", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "ユーザーが見つからない", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/users/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserResponse> getUserById(
        @Parameter(name = "id", description = "ユーザーID", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    /**
     * GET /api/users : ユーザー一覧取得（管理者専用）
     * ページネーション付きでユーザー一覧を取得する。管理者権限が必要。
     *
     * @param page ページ番号（0始まり） (optional, default to 0)
     * @param size 1ページあたりの件数 (optional, default to 20)
     * @param sort ソート条件（例: createdAt,desc） (optional)
     * @return 取得成功 (status code 200)
     *         or 未認証 (status code 401)
     *         or 権限不足（管理者権限が必要） (status code 403)
     */
    @Operation(
        operationId = "getUsers",
        summary = "ユーザー一覧取得（管理者専用）",
        description = "ページネーション付きでユーザー一覧を取得する。管理者権限が必要。",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "取得成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PageUserResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "未認証", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "権限不足（管理者権限が必要）", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/users",
        produces = { "application/json" }
    )
    
    ResponseEntity<PageUserResponse> getUsers(
        @Min(0) @Parameter(name = "page", description = "ページ番号（0始まり）", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(1) @Parameter(name = "size", description = "1ページあたりの件数", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size,
        @Parameter(name = "sort", description = "ソート条件（例: createdAt,desc）", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) String sort
    );


    /**
     * PUT /api/users/me : ログインユーザー情報更新
     * 認証済みユーザー自身の情報を更新する
     *
     * @param updateUserRequest  (required)
     * @return 更新成功 (status code 200)
     *         or バリデーションエラー (status code 400)
     *         or 未認証 (status code 401)
     */
    @Operation(
        operationId = "updateCurrentUser",
        summary = "ログインユーザー情報更新",
        description = "認証済みユーザー自身の情報を更新する",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "更新成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "バリデーションエラー", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "未認証", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/users/me",
        produces = { "application/json" },
        consumes = "application/json"
    )
    
    ResponseEntity<UserResponse> updateCurrentUser(
        @Parameter(name = "UpdateUserRequest", description = "", required = true) @Valid @RequestBody UpdateUserRequest updateUserRequest
    );

}